feat: added logic for Beverage controller

+++ Brewview/src/main/java/com/dto/BeverageRequestDTO.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public class BeverageRequestDTO {
 
     @NotBlank(message = "Beverage name cannot be blank")
@@ -17,7 +19,7 @@ public class BeverageRequestDTO {
 
     @NotNull(message = "ABV cannot be null")
     @Min(value = 0, message = "ABV must be 0 or greater")
-    private int abv;
+    private BigDecimal abv;
 
     @NotBlank(message = "Description cannot be blank")
     @Size(min = 25, max = 254, message = "The Beverage Description must be between 25 and 254 characters")
@@ -30,7 +32,7 @@ public class BeverageRequestDTO {
     public BeverageRequestDTO() {
     }
 
-    public BeverageRequestDTO(String beverageName, String type, int abv, String description, String imageUrl) {
+    public BeverageRequestDTO(String beverageName, String type, BigDecimal abv, String description, String imageUrl) {
         this.beverageName = beverageName;
         this.type = type;
         this.abv = abv;
@@ -54,11 +56,11 @@ public class BeverageRequestDTO {
         this.type = type;
     }
 
-    public int getAbv() {
+    public BigDecimal getAbv() {
         return abv;
     }
 
-    public void setAbv(int abv) {
+    public void setAbv(BigDecimal abv) {
         this.abv = abv;
     }
